Pattern Matching FTW!
I have no problem with starting a solution with shameless greeen.
I spent 20m or so trying to think of ways to use all kinds of
fancy functions to break this apart. Right now, though, pattern matching
is all that is needed. If there was some form of infinite options,
then this approach would not work. I'm okay with it as is.