Name Branded Elohim
Begin at the beginning of the beginning. The first thing we must learn is the name of who spoke this all into being. Notice that every message mentioned in this fake farm society goes against what seems to be the most obvious truths that we inately know from experiences, common senses, and best selling books. The name begins it all.