It's all about Content
I'm learning that a program is only as good as what needs to be done.
Code by itself is not valuable if it doesn't get the content from point
A to B. Both 'A' and 'B' need to be defined. For me, it's a few formats
as these fixture files demonstrate.